app/requirements.txt
Django==2.0
djangorestframework==3.7.4
django-filter==1.1.0
mysqlclient==1.3.12
py-auth0-jwt-rest==0.1.2
PyJWT==1.4.2
PyMySQL==0.7.9
python-pstore==0.8
pytz==2017.3
